fix: 修复请求日志流式判断/token提取/耗时计算，优化前端交互
- cmd/agw/main.go: isStream 增加 body 中 stream 字段检测
- internal/proxy/engine.go: ForwardStream lastChunk 改为累积最后 64KB 确保 usage 不丢失
- internal/proxy/engine.go: upstream_latency_ms 改为从上游响应头提取(openai-processing-ms 等)
- internal/stats: AvgLatencyMs 从 int 改为 float64 修复 SQL AVG 扫描类型不匹配
- web: token 展示用千分位分隔，详情 token 同步
- web: 模型链路模型名改为 n-tag 可点击复制，前后不同颜色区分
- web: 修复搜索框 clear 事件 v-model 未更新导致搜索条件残留 bug
- web: 搜索按钮增加 loading 状态防重复点击
- web: 重置按钮改为当天时间区间而非清空
- web: SystemLogs 原始数据背景改为黑底白字

+++ internal/stats/async_writer.go
@@ -221,7 +221,7 @@ func (c *TodayCounters) Increment(success bool, latencyMs int, tokens int) {
 }
 
 // Snapshot 获取当前快照
-func (c *TodayCounters) Snapshot() (date string, total, success, fail int64, avgLatencyMs int, tokens int64) {
+func (c *TodayCounters) Snapshot() (date string, total, success, fail int64, avgLatencyMs float64, tokens int64) {
 	date = c.date.Load().(string)
 	total = c.totalRequests.Load()
 	success = c.successRequests.Load()
@@ -229,7 +229,7 @@ func (c *TodayCounters) Snapshot() (date string, total, success, fail int64, avg
 	tokens = c.totalTokens.Load()
 	count := c.requestCount.Load()
 	if count > 0 {
-		avgLatencyMs = int(c.totalLatencyMs.Load() / count)
+		avgLatencyMs = float64(c.totalLatencyMs.Load()) / float64(count)
 	}
 	return
 }
